pyproject.toml
[project]
name = "fontis-ai-voice-agent"
version = "0.1.0"
description = "FastAPI middleware for Vapi-Fontis Water API integration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.32.0,<0.33.0",
    "pydantic>=2.9.0,<3.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "httpx>=0.27.0,<0.28.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "python-multipart>=0.0.9,<1.0.0",
    "gspread>=6.0.0,<7.0.0",
    "google-auth>=2.0.0,<3.0.0",
    "twilio>=9.0.0,<10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.7.0",
    "mypy>=1.8.0",
    "pytest-httpx>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

